pyproject.toml
[tool.poetry]
name = "ccp"
version = "0.1.0"
description = "Club Campestre de Pereira Tenis CLI Utilities"
authors = [
  "Luis Rueda <userlerueda@gmail.com>"
]
include = ["LICENSE"]

[tool.poetry.dependencies]
python = "^3.9"
daiquiri = "^3.2.1"
tabulate = "^0.8.10"
click = "^8.1.3"
tsw = { git = "https://github.com/userlerueda/tournamentsw.git", branch = "main" }
utr = { git = "https://github.com/userlerueda/utr.git", branch = "main" }
firebase-admin = "^5.4.0"
pydash = "^5.1.1"
pandas = "^1.5.0"
numpy = "^1.23.3"
Django = "^4.1.1"
djangorestframework = "^3.14.0"
Markdown = "^3.4.1"
django-filter = "^22.1"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
black = "^22.8.0"
pytest = "^7.1.3"
ipython = "^8.5.0"
pytest-cov = "^3.0.0"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | \.docker
    | \*.egg-info
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
profile = "black"
line_length = 79
force_grid_wrap = 0
multi_line_output = 3
use_parentheses = true
include_trailing_comma = true
ensure_newline_before_comments = true

[tool.poetry.scripts]
ccp = "ccp.main:cli"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"